FD helpers into dedicated file
